#7 DB에 입력한 내용 삽입 백엔드 수정
# Conflicts:
#	backend/.idea/misc.xml
#	backend/.idea/modules.xml

+++ backend/bookbook/src/main/java/com/bookbook/domain/rent/dto/RentResponseDto.java
@@ -3,7 +3,7 @@ package com.bookbook.domain.rent.dto;
 import java.time.LocalDateTime;
 
 public record RentResponseDto(
-        int lender_user_id,
+        Long lender_user_id,
         String bookCondition,
         String bookImage,
         String address,
